Continued model calculation methods
git-svn-id: file:///home/simon/tmp/svnmigrate/svn/fsfibu2@195 200f3aa7-6921-4cde-8511-3937f5af78e3

+++ src/fs/fibu2/view/model/BilancialInformation.java
@@ -19,17 +19,17 @@ public class BilancialInformation {
 	/**
 	 * The overall sum of all entries
 	 */
-	public int overallSum = 0;
+	public float overallSum = 0;
 	
 	/**
 	 * Sums of entries in certain categories
 	 */
-	public HashMap<Category, Integer> categorySums = new HashMap<Category, Integer>();
+	public HashMap<Category, Float> categorySums = new HashMap<Category, Float>();
 	
 	/**
 	 * Sums of entries using certain accounts
 	 */
-	public HashMap<Account, Integer> accountSums = new HashMap<Account, Integer>();
+	public HashMap<Account, Float> accountSums = new HashMap<Account, Float>();
 	
 	
 }
